i am getting error com.fasterxml.jackson.databind.exc.InvalidDefinitionException: Cannot construct instance of `com.github.innovationforge.model.User` (no Creators, like default constructor, exist): cannot deserialize from Object value (no delegate- or property-based Creator)
The error message you’re seeing is due to the fact that Jackson, the library you’re using to parse the JSON, requires a default constructor (a constructor with no arguments) to instantiate the User object before it can populate the fields with the parsed values.

However, in your User class, you only have an all-args constructor (provided by the @AllArgsConstructor annotation from Lombok). This is why Jackson is complaining about not being able to find a “creator” (i.e., a constructor).

To fix this issue, you can add a default constructor to your User class. If you want to keep your class immutable, you can use the @NoArgsConstructor and @AllArgsConstructor annotations together with @JsonProperty on the fields. Here’s how you can do it: