AtomicReference is a class in the java.util.concurrent.atomic package in Java. It provides an object reference that may be updated atomically1. This means that when multiple threads try to access or modify the object reference, these operations are done in a thread-safe manner2.
Here are some of the key methods provided by AtomicReference:

get(): Gets the current value1.
set(newValue): Sets the value to the given value1.
compareAndSet(expect, update): Atomically sets the value to the given updated value if the current value == the expected value1.
getAndSet(newValue): Atomically sets to the given value and returns the old value1.
lazySet(newValue): Eventually sets to the given value1.
These methods help in performing atomic operations on the underlying object reference variable3. This is particularly useful in multi-threaded environments where you want to ensure the integrity of the object reference when accessed or modified by multiple threads2.